refactor: reorganize fastapi backend structure
