+++ .github/pull_request_template.md
@@ -1,0 +1,17 @@
+## Summary
+- 
+
+## Changes
+- 
+
+## Testing
+- [ ] `python -m compileall project/backend`
+- [ ] Manual API check
+- [ ] Not run
+
+## Impact
+- Affected areas:
+- Rollback plan:
+
+## Notes
+-
