+++ project/backend/app/__init__.py
@@ -1,0 +1,3 @@
+from project.backend.app.factory import create_app
+
+__all__ = ["create_app"]
